requirements.txt
mitmproxy>=10.0.0
click>=8.0.0
colorama>=0.4.4
PyYAML>=6.0
